chore: update folder naming (#2541)
