+++ camel/data_collectors/base.py
