+++ docs/mintlify/key_modules/runtimes.mdx
@@ -101,7 +101,7 @@ long as they follow the `FunctionTool` type defined by CAMEL (see [tools](
 ./tools.md) 
 module document for more details.)
 ```
-from camel.runtime import RemoteHttpRuntime
+from camel.runtimes import RemoteHttpRuntime
 from camel.toolkits import MathToolkit
 
 if __name__ == "__main__":
