+++ docs/mintlify/reference/camel.data_collectors.alpaca_collector.mdx
@@ -1,0 +1,64 @@
+<a id="camel.data_collectors.alpaca_collector"></a>
+
+<a id="camel.data_collectors.alpaca_collector.AlpacaDataCollector"></a>
+
+## AlpacaDataCollector
+
+```python
+class AlpacaDataCollector(BaseDataCollector):
+```
+
+<a id="camel.data_collectors.alpaca_collector.AlpacaDataCollector.__init__"></a>
+
+### __init__
+
+```python
+def __init__(self):
+```
+
+<a id="camel.data_collectors.alpaca_collector.AlpacaDataCollector.record"></a>
+
+### record
+
+```python
+def record(self, agent: Union[List[ChatAgent], ChatAgent]):
+```
+
+Inject an agent into the data collector.
+
+**Parameters:**
+
+- **agent** (Union[List[ChatAgent], ChatAgent]): The agent to inject.
+
+<a id="camel.data_collectors.alpaca_collector.AlpacaDataCollector.convert"></a>
+
+### convert
+
+```python
+def convert(self):
+```
+
+Convert the collected data into a dictionary.
+
+<a id="camel.data_collectors.alpaca_collector.AlpacaDataCollector.llm_convert"></a>
+
+### llm_convert
+
+```python
+def llm_convert(
+    self,
+    converter: Optional[OpenAISchemaConverter] = None,
+    prompt: Optional[str] = None
+):
+```
+
+Convert collected data using an LLM schema converter.
+
+**Parameters:**
+
+- **converter** (Optional[OpenAISchemaConverter], optional): The converter to use. (default: :obj:`OpenAISchemaConverter`)
+- **prompt** (Optional[str], optional): Prompt to guide the conversion. (default: :obj:`DEFAULT_CONVERTER_PROMPTS`)
+
+**Returns:**
+
+  Dict[str, str]: The converted data.
