+++ examples/runtimes/remote_http_runtime.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
-from camel.runtime import RemoteHttpRuntime
+from camel.runtimes import RemoteHttpRuntime
 from camel.toolkits import MathToolkit
 
 if __name__ == "__main__":
