+++ examples/runtimes/ubuntu_docker_runtime/manage_camel_docker.sh
