+++ test/runtimes/test_code_execution_with_llm_guard_runtime.py
@@ -16,7 +16,7 @@ from camel.agents import ChatAgent
 from camel.configs import ChatGPTConfig
 from camel.messages import BaseMessage
 from camel.models import ModelFactory
-from camel.runtime import LLMGuardRuntime
+from camel.runtimes import LLMGuardRuntime
 from camel.toolkits.code_execution import CodeExecutionToolkit
 from camel.types import ModelPlatformType, ModelType
